[INTERNAL] sap.ui.model.odata.v4.ODataListBinding#refreshSingle: Support refreshing a kept-alive context
PS1: OPA tests
PS2: Integration tests
PS3: Review comments
PS4: PoC coding + OPA improvement
PS5-6: Review comments
PS7: PoC coding removed, TDD _Cache#replaceElement
PS8: TDD _Cache#refreshSingle
PS9: TDD + Refactoring _Cache#refreshSingleWithRemove
PS10: Review comments
PS11: TDD ODLB#fetchCache
PS12: TDD ODLB new event 'keepAliveContextDeleted'
PS13: TDD ODLD#refreshSingle
PS14: TDD review comment (_Cache#getElementIndex)
PS15-16: Review comments
PS17: JSDoc for Context#refresh & ODLB#refreshSingle
PS18-22: Review comments
PS23: keepAliveContextDestroyed event -> fnOnBeforeDestroy callback
PS24-26: Review comments
PS27: PP KTH/URE: review comments (partly)
PS28: Rebase
PS29: Review comments
PS30: Rebase
PS31: Review comments
PS32: Voter failure
PS33: Review comments TCH OPA + application + MIT + ODM.integration + iMaxLength fix
PS34: Review comments after sync meeting
PS35: Review comments TCH + remove deletion of oContext.fnOnRemove
PS36: Review comments after sync meeting
PS37: Coding adapted after sync with TCH (B2->A1, context leaves collection)
PS38-39: Integration test + Fix of $top=1 problem
PS40-44: Review comments
PS45: Rebase
PS46-52: Review comments

JIRA: CPOUI5ODATAV4-429
JIRA: CPOUI5ODATAV4-430
JIRA: CPOUI5ODATAV4-431
JIRA: CPOUI5ODATAV4-432
JIRA: CPOUI5ODATAV4-433
JIRA: CPOUI5ODATAV4-434
JIRA: CPOUI5ODATAV4-435
JIRA: CPOUI5ODATAV4-436
JIRA: CPOUI5ODATAV4-437
JIRA: CPOUI5ODATAV4-438
JIRA: CPOUI5ODATAV4-491

+++ src/sap.ui.core/src/sap/ui/model/odata/v4/ODataListBinding.js
@@ -229,7 +229,7 @@ sap.ui.define([
 				var sContextPath, i, sPredicate, sResolvedPath, i$skipIndex;
 
 				if (oContext.isKeepAlive()) {
-					oContext.setKeepAlive(false); // ensure that it is destroyed later
+					oContext.resetKeepAlive(); // ensure that it is destroyed later
 				}
 				if (oContext.created()) {
 					// happens only for a created context that is not transient anymore
@@ -834,7 +834,7 @@ sap.ui.define([
 				this.aContexts[i] = oContext;
 			}
 		}
-		// destroy previous contexts which are not reused
+		// destroy previous contexts which are not reused or kept-alive
 		if (Object.keys(this.mPreviousContextsByPath).length) {
 			oModel.addPrerenderingTask(this.destroyPreviousContexts.bind(this));
 		}
@@ -1079,6 +1079,7 @@ sap.ui.define([
 	ODataListBinding.prototype.fetchCache = function () {
 		var oOldCache = this.oCache,
 			sBindingPath = this.oModel.resolve(this.sPath, this.oContext),
+			bKeptDataAdded,
 			that = this;
 
 		asODataParentBinding.prototype.fetchCache.apply(this, arguments);
@@ -1087,13 +1088,17 @@ sap.ui.define([
 			this.oCachePromise.then(function (oCache) {
 				Object.keys(that.mPreviousContextsByPath).forEach(function (sPath) {
 					var oContext = that.mPreviousContextsByPath[sPath];
+
 					if (oContext.isKeepAlive()) {
+						// either absolute or $$ownRequest === true
+						// ->has own cache, see #checkKeepAlive
 						oCache.addKeptElement(
 							oOldCache.getValue(_Helper.getRelativePath(sPath, sBindingPath)));
 						oContext.checkUpdate(); // add change listeners
+						bKeptDataAdded = true;
 					}
 				});
-				if (oCache && oCache.hasChangeListeners()) {
+				if (bKeptDataAdded) {
 					oCache.setLateQueryOptions(oOldCache.getLateQueryOptions());
 				}
 			}); // no catch; if the new cache can't be created, data for kept-alive contexts is lost
@@ -1743,7 +1748,8 @@ sap.ui.define([
 		var that = this;
 
 		return this.oModel.getDependentBindings(this).filter(function (oDependentBinding) {
-			return !(oDependentBinding.oContext.getPath() in that.mPreviousContextsByPath);
+			return oDependentBinding.oContext.isKeepAlive()
+				|| !(oDependentBinding.oContext.getPath() in that.mPreviousContextsByPath);
 		});
 	};
 
@@ -2197,6 +2203,9 @@ sap.ui.define([
 	 *   see {@link sap.ui.model.odata.v4.ODataListBinding#filter}; a removed context is
 	 *   destroyed, see {@link sap.ui.model.Context#destroy}.
 	 *   Supported since 1.55.0
+	 *
+	 *   A removed context is destroyed unless it is kept alive
+	 *   (see {@link sap.ui.model.odata.v4.Context#isKeepAlive}) and still exists on the server.
 	 * @returns {sap.ui.base.SyncPromise}
 	 *   A promise which resolves with the entity when the entity is updated in the
 	 *   cache, or <code>undefined</code> if <code>bAllowRemoval</code> is set to true.
@@ -2206,7 +2215,8 @@ sap.ui.define([
 	 * @private
 	 */
 	ODataListBinding.prototype.refreshSingle = function (oContext, oGroupLock, bAllowRemoval) {
-		var sResourcePathPrefix = oContext.getPath().slice(1),
+		var sContextPath = oContext.getPath(),
+			sResourcePathPrefix = sContextPath.slice(1),
 			that = this;
 
 		if (oContext === this.oHeaderContext) {
@@ -2215,8 +2225,10 @@ sap.ui.define([
 
 		return this.withCache(function (oCache, sPath, oBinding) {
 			var bDataRequested = false,
-				aPromises = [],
-				bRemoved = false;
+				bDestroyed = false,
+				bKeepAlive = oContext.isKeepAlive(),
+				sPredicate = _Helper.getRelativePath(sContextPath, that.oHeaderContext.getPath()),
+				aPromises = [];
 
 			function fireDataReceived(oData) {
 				if (bDataRequested) {
@@ -2229,37 +2241,61 @@ sap.ui.define([
 				that.fireDataRequested();
 			}
 
-			function onRemove() {
-				var i, iIndex;
+			/**
+			 * Removes this context from the list bindings collection as it no longer matches the
+			 * filter criteria, see
+			 * {@link sap.ui.model.odata.v4.lib._Cache#refreshSingleWithRemove}.
+			 *
+			 * @param {boolean} bStillAlive
+			 *   If <code>false</code>, the context does not match the filter criteria and, if the
+			 *   context is kept-alive, the entity it points to no longer exists. If
+			 *   <code>true</code>, the context is kept-alive and the entity it points to still
+			 *   exists. In this case the context must not be destroyed.
+			 */
+			function onRemove(bStillAlive) {
+				var i,
+					iIndex = oContext.getModelIndex();
 
 				if (oContext.created()) {
 					that.destroyCreated(oContext);
+					bDestroyed = true;
 				} else {
-					iIndex = oContext.getModelIndex();
-					that.aContexts.splice(iIndex, 1);
-					for (i = iIndex; i < that.aContexts.length; i += 1) {
-						if (that.aContexts[i]) {
-							that.aContexts[i].iIndex -= 1;
+					if (iIndex === undefined) { // -> bStillAlive === false
+						delete that.mPreviousContextsByPath[sContextPath];
+					} else {
+						that.aContexts.splice(iIndex, 1);
+						that.iMaxLength -= 1; // this doesn't change Infinity
+						for (i = iIndex; i < that.aContexts.length; i += 1) {
+							if (that.aContexts[i]) {
+								that.aContexts[i].iIndex -= 1;
+							}
+						}
+						if (bStillAlive) {
+							that.mPreviousContextsByPath[sContextPath] = oContext;
 						}
 					}
-					oContext.destroy();
-					that.iMaxLength -= 1; // this doesn't change Infinity
+
+					if (!bStillAlive) {
+						bDestroyed = true;
+						oContext.destroy();
+					}
 				}
-				bRemoved = true;
-				that._fireChange({reason : ChangeReason.Remove});
+				if (iIndex !== undefined) {
+					that._fireChange({reason : ChangeReason.Remove});
+				}
 			}
 
 			aPromises.push(
 				(bAllowRemoval
 					? oCache.refreshSingleWithRemove(oGroupLock, sPath, oContext.getModelIndex(),
-						fireDataRequested, onRemove)
-					: oCache.refreshSingle(oGroupLock, sPath, oContext.getModelIndex(),
-						fireDataRequested))
+						sPredicate, bKeepAlive, fireDataRequested, onRemove)
+					: oCache.refreshSingle(oGroupLock, sPath, oContext.getModelIndex(), sPredicate,
+						bKeepAlive, fireDataRequested))
 				.then(function (oEntity) {
 					var aUpdatePromises = [];
 
 					fireDataReceived({data : {}});
-					if (!bRemoved) { // do not update removed context
+					if (!bDestroyed) { // do not update destroyed context
 						aUpdatePromises.push(oContext.checkUpdate());
 						if (bAllowRemoval) {
 							aUpdatePromises.push(
@@ -2500,7 +2536,7 @@ sap.ui.define([
 		function reset(oContext) {
 			// do not call it always, it throws an exception on a relative binding w/o $$ownRequest
 			if (oContext.isKeepAlive()) {
-				oContext.setKeepAlive(false);
+				oContext.resetKeepAlive();
 			}
 		}
